Added User Roles

+++ app/Http/Controllers/API/V1/UserController.php
@@ -6,9 +6,10 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use App\Http\Requests\API\V1\IndexRequest;
-use Illuminate\Validation\Rules\Password;
-use Hash;
-use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\API\V1\User\StoreUserRequest;
+use App\Http\Requests\API\V1\User\UpdateUserRequest;
+
 
 class UserController extends ApiController
 {
@@ -18,6 +19,7 @@ class UserController extends ApiController
     public function index(IndexRequest $request)
     {
 
+        Gate::authorize('viewAny', User::class);
         return $this->sendResponseWithPaginatedData(
             User::when($request->search, function (Builder $query, string $search) {
                 $query->where(function (Builder $q) use($search) {
@@ -36,16 +38,8 @@ class UserController extends ApiController
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StoreUserRequest $request)
     {
-        $request->validate([
-            "name" => ["required", "string"],
-            "email" => ["required", "email", "unique:users,email"],
-            "password" => ["required", $this->passwordCriteria(), "confirmed"],
-            "password_confirmation" => ["required"],
-        ]);
-
-        $request->merge(["password" => Hash::make($request->password)]);
         $user = User::create($request->all());
         return $this->sendResponse(__("Created Successfully"),$user);
 
@@ -56,22 +50,15 @@ class UserController extends ApiController
      */
     public function show(User $user)
     {
+        Gate::authorize('view', $user);
         return $this->sendResponse(__("Fetched Successfully"),$user);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, User $user)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        $request->validate([
-            "name" => ["required", "string"],
-            "email" => ["required", "email", Rule::unique('users')->ignore($user->id)],
-            "old_password" => ["required_with:password", "current_password"],
-            "password" => [ $this->passwordCriteria(), "confirmed"],
-            "password_confirmation" => ["required_with:password"],
-        ]);
-        $request->merge(["password" => Hash::make($request->password)]);
         $user->update($request->all());
         return $this->sendResponse(__("Updated Successfully"),$user);
     }
@@ -81,15 +68,9 @@ class UserController extends ApiController
      */
     public function destroy(User $user)
     {
+        Gate::authorize('delete', $user);
         $user->delete();
         return $this->sendResponse(__("Deleted Successfully"));
     }
 
-    /**
-     * get Password criteria.
-     */
-    public function passwordCriteria() : Password
-    {
-        return Password::min(8)->letters();
-    }
 }
